(refactor) Implemented production/development key strategy
Snippets in Slack channel for the files you need to make
