+++ README.md
@@ -1,16 +1,15 @@
-# Project Name
+# LePlonger
 
-> Pithy project description
+> LePlonger is an application that provides weather services and wave data analysis for scuba divers. 
 
 ## Team
-
   - John Iaconnis
   - Colby Duhon
   - Aly Tamboura
   - Jackson Carter
 
 ## Table of Contents
-
+1. [Demo](#Demo)
 1. [Usage](#Usage)
 1. [Requirements](#requirements)
 1. [Development](#development)
@@ -18,18 +17,17 @@
     1. [Tasks](#tasks)
 1. [Roadmap](#roadmap)
 1. [Contributing](#contributing)
+## Demo
+Coming Soon!
 
 ## Usage
 
-> Some usage instructions
+Signing up for an account allows a user to add new dive sites to the ever growing list of sites along the Pacific Coast. Visitors to the site can simply need to select a dive site they're interested in and click 'Check it out!' to read reviews and get the current diving conditions.
 
 ## Requirements
 
 - Node 6.4.x
-- Redis 2.6.x
-- Postgresql 9.1.x
-- etc
-- etc
+- TBD
 
 ## Development
 
@@ -40,13 +38,11 @@ From within the root directory:
 ```sh
 npm install -g bower
 npm install
-bower install
 ```
 
 ### Roadmap
 
 View the project roadmap [here](LINK_TO_DOC)
-
 
 ## Contributing
