+++ config/index.js
@@ -1,0 +1,13 @@
+const env = process.env.NODE_ENV || 'development';
+
+let config;
+
+if (env === 'development') {
+  config = require('./keys.js');
+}
+
+if (env === 'production') {
+  config = require('./config.prod.js');
+}
+
+module.exports = config;
