+++ server/models/index.js
@@ -1,6 +1,6 @@
 const connection = require('../db');
 const request = require('request');
-const Api = require('../../keys');
+const Api = require('../../config/index.js');
 const Promise = require('bluebird');
 const axios = require('axios');
 const sha1 = require('sha1');
@@ -97,7 +97,7 @@ module.exports = {
  //uncomment url for actual use, disabled so we don't hit api limit
     get: (req, res) => {
       const location = `${req.body.location.lat},${req.body.location.lng}`;
-      // const url = `http://api.wunderground.com/api/${Api.weatherUnderground}/geolookup/conditions/q/${location}.json`;
+      const url = `http://api.wunderground.com/api/${Api.weatherUnderground}/geolookup/conditions/q/${location}.json`;
 
       axios.get(url)
         .then( (result) => {
